+++ src/fixtures/maskReference.js
@@ -1,0 +1,16 @@
+import image from './image';
+import url from './url';
+
+export default {
+    valid: [
+        'image(url(mask.png), skyblue), linear-gradient(90deg, transparent, blue)',
+        'none',
+        ...url.valid,
+        ...image.valid,        
+    ],
+    invalid: [
+        ...url.invalid,
+        ...image.invalid,
+        'abc',
+    ],
+};